Rename new feature copyright - updateheader
