+++ pontos/updateheader/templates/AGPL-3.0-or-later/template.js
