+++ pontos/updateheader/templates/GPL-2.0-only/template.c
