+++ pontos/updateheader/templates/GPL-2.0-or-later/template.bash
