+++ pontos/updateheader/templates/GPL-3.0-or-later/template.xml
